ensure job manager contextualization
